fix: range error upon decreasing page count, added tests, bumped version to 1.0.10, updated readme and changelog
