+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: expandable_page_view
 description: A PageView widget adjusting it's height to currently displayed page
-version: 1.0.9
+version: 1.0.10
 homepage: https://github.com/Limbou/expandable_page_view
 
 environment:
