+++ test/helpers/page_view_height.dart
@@ -1,0 +1,6 @@
+import 'package:expandable_page_view/src/expandable_page_view.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+extension PageViewHeight on WidgetTester {
+  double get pageViewHeight => getSize(find.byType(ExpandablePageView)).height;
+}
