+++ test/helpers/pump_app.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+extension PumpApp on WidgetTester {
+  Future<void> pumpApp(Widget widget) async {
+    await pumpWidget(
+      MaterialApp(
+        home: Scaffold(body: widget),
+      ),
+    );
+    await pumpAndSettle();
+  }
+}
